requirements.txt
torch==2.0.1
transformers==4.38.0
sentencepiece==0.1.99
accelerate==0.23.0
peft==0.4.0
bitsandbytes==0.41.0
scikit-learn==1.2.2
einops==0.6.1
deepspeed==0.9.5
pillow
shortuuid
numpy<2.0.0